Extract codegen and infer into their own crates
