+++ Cargo.toml
@@ -3,5 +3,7 @@
 members = [
     "crates/crochet",
     "crates/crochet_ast",
+    "crates/crochet_codegen",
+    "crates/crochet_infer",
     "crates/crochet_parser",
 ]
