+++ crates/crochet_codegen/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "crochet_codegen"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+crochet_ast = { version = "0.1.0", path = "../crochet_ast" }
+crochet_infer = { version = "0.1.0", path = "../crochet_infer" }
+swc_atoms = "0.2.11"
+swc_ecma_ast = "0.78.0"
+swc_common = "0.18.2"
+swc_ecma_codegen = "0.108.2"
+swc_ecma_transforms_react = "0.114.1"
+swc_ecma_visit = "0.64.0"
