+++ crates/crochet_infer/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "crochet_infer"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+crochet_ast = { version = "0.1.0", path = "../crochet_ast" }
+defaultmap = "0.5.0"
+itertools = "0.10.3"
+
+[dev-dependencies]
+chumsky = "0.8.0"
+crochet_parser = { version = "0.1.0", path = "../crochet_parser" }
