+++ crates/crochet_infer/src/constraint_solver.rs
