+++ crates/crochet_infer/src/infer_prog.rs
