+++ crates/crochet_infer/src/types/lit.rs
