+++ crates/crochet_infer/src/util.rs
